backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
transformers==4.37.2
torch==2.1.2
torchvision==0.16.2
pillow==10.2.0
numpy==1.26.3
qdrant-client==1.7.3
ftfy==6.1.3
regex==2023.12.25
tqdm==4.66.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
kagglehub==0.2.9
easyocr==1.7.1
opencv-python-headless==4.8.1.78
openai>=1.0.0
